feat: Show image preview for image artifacts.
I don't really like it, the code is kinda fug, the UI is mid, but I can't think with all this drilling going on in the building.
Nobody is using the app, or looking at the codebase, the world is coming to an end soon anyway; none of it really matters.